Calculate AUC for Sparse Data

This is an implementation of Bailors method for calculating AUCs with sparse sampling. It is taken from the following publication:

Nedelman, J. R., Gibiansky, E., & Lau, D. T. (1995). Applying Bailer's method for AUC confidence intervals to sparse sampling Pharmaceutical Research, 12(1), 124-128.

AUC_Bailers_method(
  conc_data = NULL,
  dsmap = list(NTIME = "NTIME", CONC = "CONC", ID = "ID")
)

Arguments

conc_data
data frame containing the sparse data

dsmap
list with names specifying the columns:

NTIME Nominal time since last dose; "NTIME" (default)

CONC Concentration data; "CONC" (default)

ID Subject ID; ("ID" (default)

Value

list with the following elements

isgood Boolean value indicating the result of the function call

AUC Mean AUC

var_AUC Variance of the AUC

msgs Sequence of strings contianing a description of any problems

obss Internal of observations

times Sequence of time corresponding to the rows of obs

r number of observations at each time point (rows correspond to rows of obs)