Calculates the Value of the Specified Objective Function

For a given set of system parameters the objective function will be calculated based on defined cohorts and variance models.

calculate_objective(parameters, cfg, estimation = TRUE)

Arguments

parameters
system parameters

cfg
ubiquity system object

estimation
boolean variable to indicate if the objective function is being called during parameter estimation

Value

If estimation is TRUE it will return the objective function value, if it is FALSE it will return a list with an element value containing the objective function value and an element named isgood that is TRUE if the objective function was successful.