Create Full Parameter Vector from Estimation Subset

Can be used to take a subset of parameters (those being estimated and returned from ' system_estimate_parameters) into a full list of system parameters.

Usage

fetch_full_parameters(pest, cfg)

Arguments

pest

list containing subset of parameters being estimated

cfg

ubiquity system object

Value

Full list of parameters with default values for the currently selected parameter set and the values in pest merged

Details

This function is used to build a full parameter set from a subset, and is normally used during parameter estimation in the observation details function when the entire parameter vector is needed to simulate the system.

The function select_set pulls out a parameter set and can optionally select only a subset for estimation:

pnames = c('Vp', 'CL')
cfg = system_select_set(cfg, "default", pnames)

The default values of this subset can be accessed in the following way:

pest = system_fetch_guess(cfg)

The estimation routines will work with this reduced parameter set, but to run simulations the full set is needed. The full values can be retrieved using the following:

parameters = fetch_full_parameters(pest, cfg) 

See also