Generates a Parameter Based on <IIV:?> in the System File

Internal function used to generate parameters based on IIV information

Usage

generate_parameter(
  SIMINT_parameters,
  SIMINT_cfg,
  SIMINT_PARAMETER_TV,
  SIMINT_IIV_VALUE,
  SIMINT_equation
)

Arguments

SIMINT_parameters

parameters vector containing the typical values

SIMINT_cfg

ubiquity system object

SIMINT_PARAMETER_TV

Typical value of the parameter in question

SIMINT_IIV_VALUE

sample from mvr distribution

SIMINT_equation

equation relating IIV and typical value to the parameter value with variability

Value

parameter value with the variability applied