Define Sample Times Around Events

When events, such as bolus doses, are applied to the system rapid changes can occur. If the system is not sampled heavily around these times, these changes may be missed in the output profiles. Based on the total duration of the sample times, extra samples can be added near these events.

sample_around(tvals, ot, sample_delta_mult = 1e-06)

Arguments

tvals
vector of event times

ot
simualtion output times

sample_delta_mult
multiplier used to control the magnitude of spacing around and following event times

Value

vector of event times and added samples

Details

For more information on setting options for population simulation see the stochastic section of the system_set_option help file.