Calculate Solution Statistics

Attempts to determine the variance/covariance matrix, confidence intervals and CV percent for a list of parameter estimates parameters. This method was taken from the ADAPT 5 User's Guide chapter 3.

solution_statistics(parameters, cfg)

Arguments

parameters
list of parameter estimates

cfg
ubiquity system object

Value

list containing information about the provided estimates

Details

The returned list has the following format:

objective - objective function value

num_observations - number of observations

degrees_of_freedom - degrees of freedom

aic - Akaike information criterion

bic - Bayesian (Schwarz) information criterion

covariance - variance covariance matrix

wls - defined for weighted least squares objective with the following elements:

jacobian - Jacobian matrix

weights - diagonal matrix of weights

error_variance - diagonal matrix of variances

ml - defined for maximum likelihood objective with the following elements:

M - Jacobian matrix with block for variance parameters

coefficient_of_variation$pname - CV percent for parameter pname

confidence_interval$lower_bound$pname - Lower bound of the confidence interval for pname

confidence_interval$upper_bound$pname - Upper bound of the confidence interval for pname

See also