Control Estimation Process

Manages the flow of parameter estimation using data specified with system_define_cohort.

Usage

system_estimate_parameters(
  cfg,
  flowctl = "plot guess",
  analysis_name = "my_analysis",
  archive_results = TRUE
)

Arguments

cfg

ubiquity system object

flowctl

string to control what the flow of the function

analysis_name

string containing the name of the analysis

archive_results

boolean variable to control whether results will be archived

Value

parameter estimates

Details

The flowctl argument can have the following values

"plot guess" return the initial guess

"estimate" perform estimation

"previous estimate as guess" load previous estimate for analysis_name and use that as the initial guess

"plot previous estimate" return the previous estimate for analysis_name