Fetch Variability Terms

Extract elements of the current variance/covariance matrix specified in the system file with <IIV:?:?> ?, <IIVCOR:?:?>?, <IIVSET:?:?> ?, <IIVCORSET:?:?>?

system_fetch_iiv(cfg, IIV1, IIV2)

Arguments

cfg
ubiquity system object

IIV1
row name of the variance/covariance matrix

IIV2
column name of the variance/covariance matrix

Value

Value from the variance/covariance matrix

See also

Examples

# \donttest{
# Creating a system file from the mab_pk example
fr = system_new(file_name        = "system.txt", 
                system_file      = "mab_pk", 
                overwrite        = TRUE, 
                output_directory = tempdir())

# Building the system 
cfg = build_system(system_file  = file.path(tempdir(), "system.txt"),
      output_directory          = file.path(tempdir(), "output"),
      temporary_directory       = tempdir())

#> #> Ubiquity: (https://ubiquity.tools)

#> #> Distribution:           package

#> #> Building the system:    /var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T//Rtmpjqdr3x/system.txt

#> #> Done 

#> #> Compiling C version of system

#> #> Loading the shared C library

#> #> System built, to fetch a new template use the following commands:

#> #>   fr = system_fetch_template(cfg, template = "Simulation")

#> #>   fr = system_fetch_template(cfg, template = "Estimation")

# Covariance term for ETACL and ETAVc
val = system_fetch_iiv(cfg, IIV1="ETACL", IIV2="ETAVc")
# }