List NCA parameters, text names and descriptions

Provides a verbose information about NCA parameters

Usage

system_nca_parameters_meta(cfg)

Arguments

cfg

ubiquity system object

Value

List with the following elements:

isgood Boolean value indicating the success of the function call.

parameters List with element names for each standard column header for NCA output. Each element name is a list with the following elements:

label Textual descriptor of the parameter.

description Verbose description of the parameter.

from Text indicating the source of the parameter (either PKNCA or ubiquity).

See also