Check NONMEM Dataset for Automatic Definitions

Checks the dataset against the information specified by system_define_cohorts_nm for validity

Usage

system_nm_check_ds(
  cfg,
  DS = "DSNAME",
  col_ID = "ID",
  col_CMT = "CMT",
  col_DV = "DV",
  col_TIME = "TIME",
  col_AMT = "AMT",
  col_RATE = "RATE",
  col_EVID = "EVID",
  col_GROUP = NULL,
  filter = NULL,
  INPUTS = NULL,
  OBS = NULL
)

Arguments

cfg

ubiquity system object

DS

Name of the dataset loaded using system_load_data

col_ID

Column of unique subject identifier

col_CMT

Compartment column

col_DV

Column with observations or ’.’ for input

col_TIME

Column with system time of each record

col_AMT

Infusion/dose amounts (these need to be in the same units specified in the system.txt file)

col_RATE

Rate of infusion or ’.’ for bolus

col_EVID

EVID (0 - observation, 1 dose)

col_GROUP

Column name to use for defining similar cohorts when generating figures.

filter

List used to filter the dataset or NULL if the whole dataset is to be used (see filter rules or nm_select_records or a description of how to use this option)

INPUTS

List mapping input information in the dataset to names used in the system.txt file

OBS

List mapping obseravation information in the dataset to names used in the system.txt file

Value

list with the following elements

"isgood" Boolean variable indicating success (TRUE) or failure (FALSE)

"mywarning" Boolean variable indicating warnings (TRUE) or no warnings (FALSE)

"dsraw" Dataframe with the filtered raw data that was used

"input_records" Rows from dsraw containing the input information

"obs_records" Rows from dsraw containing the observation information

"sids" Subject ids found in dsraw

"TSsys" system time scale used in the dataset