Sets Placeholder Content for Word Document Report

Adds or updates content to be substituted for placeholders in the specified report.

For example if you have ===HEADERLEFT=== in the header of your document and you wanted to replace it with the text "Upper left" you would do the following:

cfg = system_report_doc_set_ph(cfg, ph_content = "Upper Left" , ph_name = "HEADERLEFT", ph_location = "header")

Notice the ph_name just has HEADERLEFT and leaves off the === at the beginning and end of the string.

system_report_doc_set_ph(
  cfg,
  rptname = "default",
  ph_name = NULL,
  ph_content = NULL,
  ph_location = "body"
)

Arguments

cfg
ubiquity system object

rptname
report name initialized with system_report_init

ph_name
name of the placeholder

ph_content
content to be replaced

ph_location
location of the placeholder: "body" (default), "header", or "footer"

Value

cfg ubiquity system object with the placeholder content set