Selecting Parameter Sets

The system file can contain multiple parameterizations using the <PSET:?:?>? notation. This function provides the means for switching between these parameterizations, and (optionally) specifying a subset of parameters estimated when performing parameter estimation.

Usage

system_select_set(cfg, set_name = "default", parameter_names = NULL)

Arguments

cfg

ubiquity system object

set_name

string containing the name of the parameter set

parameter_names

list of parameter names to be estimated

Value

Ubiquity system object with the specified parameter set active

Examples

# \donttest{
# Creating a system file from the mab_pk example
fr = system_new(file_name        = "system.txt", 
                system_file      = "mab_pk", 
                overwrite        = TRUE, 
                output_directory = tempdir())

# Building the system 
cfg = build_system(system_file  = file.path(tempdir(), "system.txt"),
      output_directory          = file.path(tempdir(), "output"),
      temporary_directory       = tempdir())
#> 
#> ── Building the system: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//RtmpYM
#> → ubiquity:     https://r.ubiquity.tools
#> → Distribution: package (2.0.4)
#> → Compiling C version of system
#> → Loading the shared C library
#> ✔ System built
#> ℹ To fetch a new analysis template use `system_fetch_template`
#> ℹ For example:
#> ℹ   fr = system_fetch_template(cfg, template = "Simulation")
#> ℹ   fr = system_fetch_template(cfg, template = "Estimation")

# Selecting the default parameter set
cfg = system_select_set(cfg, "default")
# }