Actual Function Called by SI_TT_RATE

The prototype function SI_TT_RATE provides an abstract interface to this function. Based on the input from SI_TT_RATE infusion rate inputs will be updated for the current titration time.

Usage

system_set_tt_rate(
  cfg,
  rate,
  times,
  levels,
  tt_ts,
  tsinfo,
  repdose = "none",
  interval = 1,
  number = 0
)

Arguments

cfg

ubiquity system object

rate

name of the infusion rate to update(Defined in <R:?>)

times

vector of switching times relative to the current titration time (in time units defined by <R:?>)

levels

vector of infusion rates (in dosing units defined by <R:?>)

tt_ts

list of timescale values for the current titration time

tsinfo

list with timescale information for inputs (bolus, rates, etc)

repdose

"none" or "sequence"

interval

interval to repeat in the units defined in <R:?>

number

number of times to repeat

Value

ubiquity system object with the infusion rates updated.