View Information About the System

Displays information (dosing, simulation options, covariates, etc) about the system.

Usage

system_view(cfg, field = "all", verbose = FALSE)

Arguments

cfg

ubiquity system object

field

string indicating the aspect of the system to display

verbose

Boolean variable that when set to true will echo the information to the screen

Value

sequence of strings with system in formation (one line per element)

The field

"all" will show all information about the system

"parameters" summary of parameter information

"bolus" currently set bolus dosing

"rate" infusion rate dosing

"covariate" covariates

"iiv" variance/covariance information

"datasets" loaded datasets

"simulation" simulation options

"estimation" estimation options

"nca" non-compartmental analyses that have been performed

Examples

# To log and display the current system information:
# \donttest{
# Creating a system file from the mab_pk example
fr = system_new(file_name        = "system.txt", 
                system_file      = "mab_pk", 
                overwrite        = TRUE, 
                output_directory = tempdir())

# Building the system 
cfg = build_system(system_file  = file.path(tempdir(), "system.txt"),
      output_directory          = file.path(tempdir(), "output"),
      temporary_directory       = tempdir())
#> 
#> ── Building the system: /var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T//RtmpjP
#> → ubiquity:     https://r.ubiquity.tools
#> → Distribution: package (2.0.2)
#> → Compiling C version of system
#> → Loading the shared C library
#> ✔ System built
#> ℹ To fetch a new analysis template use `system_fetch_template`
#> ℹ For example:
#> ℹ   fr = system_fetch_template(cfg, template = "Simulation")
#> ℹ   fr = system_fetch_template(cfg, template = "Estimation")

  msgs = system_view(cfg, verbose=TRUE)
#> Warning: one argument not used by format ' %s, %s(%s)'
#> Warning: one argument not used by format ' %s, %s(%s)'
#> Warning: one argument not used by format ' %s, %s(%s)'
#> Warning: one argument not used by format ' %s, %s(%s)'
#> Warning: one argument not used by format ' %s, %s(%s)'
#> Warning: one argument not used by format ' '
#> →  Parameter Information
#> →  Parameter set selected:
#> →    Short Name:  default
#> →    Description: mAb in Humans
#> →  Default parameters for current set:
#> →               name |         value |           units
#> → ----------------------------------------------------
#> →                 F1 |        0.7440 |             ---
#> →                 ka |        0.2820 |           1/day
#> →                 CL |        0.2000 |           L/day
#> →                 Vc |        3.6100 |               L
#> →                 Vp |        2.7500 |               L
#> →                  Q |        0.7470 |           L/day
#> →                 MW |      140.0000 |              kD
#> →           prop_err |        0.1000 |              --
#> →            add_err |        0.1000 |           ng/ml
#> → ----------------------------------------------------
#> →  
#> →  Bolus dosing details 
#> →      field |      values |    scaling |      units
#> → --------------------------------------------------
#> →      times |      0 7 14 |          1 |       days
#> →         Cc |       0 0 0 |  1e3/MW/Vc |         mg
#> →         At |     400 0 0 |     1e3/MW |         mg
#> → --------------------------------------------------
#> →  Infusion rate details 
#> →      Rate  |      field |     values |    scaling |      units
#> → -----------------------------------------------------------------
#> →       Dinf |       time |       0 30 |    1/60/24 |        min
#> →            |     levels |        0 0 | 60*24*1e3/MW |     mg/min
#> → -----------------------------------------------------------------
#> →  
#> →  Covariate details
#> →  Covariate |      field |     values |      units
#> → --------------------------------------------------
#> →       DOSE |       time |          0 |        day
#> →     (step) |     levels |        400 |         mg
#> →         WT |       time |          0 |        day
#> →     (step) |     levels |         60 |         kg
#> → --------------------------------------------------
#> →  
#> →  IIV details
#> →  IIV/Parameter set:
#> →    Short Name:  default 
#> →  Variance/covariance matrix
#> →  
#> →                                ETAka             ETACL             ETAVc             ETAVp              ETAQ
#> →              ETAka            0.4160                 0                 0                 0                 0
#> →              ETACL                 0            0.0988            0.0786            0.0619                 0
#> →              ETAVc                 0            0.0786            0.1160            0.0377                 0
#> →              ETAVp                 0            0.0619            0.0377            0.0789                 0
#> →               ETAQ                 0                 0                 0                 0            0.6990
#> →  
#> →  On parameters
#> →           Q,       ETAQ(LN)
#> →          CL,      ETACL(LN)
#> →          ka,      ETAka(LN)
#> →          Vc,      ETAVc(LN)
#> →          Vp,      ETAVp(LN)
#> →  
#> →  integrate_with          c-file
#> →  output_times            min = 0; max = 10.0; length = 101  
#> →  No datasets loaded
#> →  
#> → Estimation details 
#> →  Parameter set:          default
#> →  Parameters estimated:   F1, ka, CL, Vc, Vp, Q, MW, prop_err, add_err
#> →  objective_type          ml
#> →  observation_function    system_od_general
#> →  No cohort information found
#> → No NCA has been performed
# }