Fetch Ubiquity Workshop Sections

With the ubiquity package this function can be used to fetch example files for different sections of the workshop.

Usage

workshop_fetch(
  section = "Simulation",
  overwrite = FALSE,
  copy_files = TRUE,
  output_directory = getwd()
)

Arguments

section

Name of the section of workshop to retrieve ("Simulation")

overwrite

if TRUE the new workshop files will overwrite any existing files present (FALSE)

copy_files

if TRUE the files will be written to the output_directory, if FALSE only the names and locations of the files will be returned (TRUE)

output_directory

directory where workshop files will be placed (getwd())

Value

list

Details

Valid sections are "Simulation", "Estimation", "In Vitro", "Titration" "Reporting", and "NCA"

Examples

# \donttest{
workshop_fetch("Estimation", output_directory=tempdir(), overwrite=TRUE)
#> → Creating file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmprz9swW/analysis_parent.r
#> → Creating file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmprz9swW/analysis_parent_metabolite.r
#> → Creating file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmprz9swW/analysis_parent_metabolite_global.r
#> → Creating file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmprz9swW/analysis_parent_metabolite_nm_data.r
#> → Creating file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmprz9swW/system.txt
#> → Creating file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmprz9swW/pm_data.csv
#> → Creating file: /var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T//Rtmprz9swW/nm_data.csv
#> $sources
#> [1] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/Rtmp5tPqMe/temp_libpathb6d24aaacedb/ubiquity/ubinc/scripts/analysis_parent.r"                   
#> [2] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/Rtmp5tPqMe/temp_libpathb6d24aaacedb/ubiquity/ubinc/scripts/analysis_parent_metabolite.r"        
#> [3] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/Rtmp5tPqMe/temp_libpathb6d24aaacedb/ubiquity/ubinc/scripts/analysis_parent_metabolite_global.r" 
#> [4] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/Rtmp5tPqMe/temp_libpathb6d24aaacedb/ubiquity/ubinc/scripts/analysis_parent_metabolite_nm_data.r"
#> [5] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/Rtmp5tPqMe/temp_libpathb6d24aaacedb/ubiquity/ubinc/systems/system-adapt.txt"                    
#> [6] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/Rtmp5tPqMe/temp_libpathb6d24aaacedb/ubiquity/ubinc/csv/pm_data.csv"                             
#> [7] "/private/var/folders/pq/7srbf_fx3rd3k706hgxkg61r0000gp/T/Rtmp5tPqMe/temp_libpathb6d24aaacedb/ubiquity/ubinc/csv/nm_data.csv"                             
#> 
#> $destinations
#> [1] "analysis_parent.r"                   
#> [2] "analysis_parent_metabolite.r"        
#> [3] "analysis_parent_metabolite_global.r" 
#> [4] "analysis_parent_metabolite_nm_data.r"
#> [5] "system.txt"                          
#> [6] "pm_data.csv"                         
#> [7] "nm_data.csv"                         
#> 
#> $write_file
#> [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> 
#> $isgood
#> [1] TRUE
#> 
# }