Automatic NCA

Performs NCA in an automated fashion

Usage

system_nca_run(
  cfg,
  dsname = "PKDS",
  dscale = 1,
  NCA_options = NULL,
  NCA_min = 4,
  analysis_name = "analysis",
  dsfilter = NULL,
  extrap_C0 = TRUE,
  extrap_N = 2,
  sparse = FALSE,
  dsmap = list(TIME = "TIME", NTIME = "NTIME", CONC = "CONC", DOSE = "DOSE", ID = "ID",
    ROUTE = "ROUTE", DOSENUM = NULL, BACKEXTRAP = NULL, SPARSEGROUP = NULL),
  dsinc = NULL
)

Arguments

cfg

ubiquity system object

dsname

name of dataset loaded with (system_load_data)

dscale

factor to multiply the dose to get it into the same units as concentration (default 1): if you are dosing in mg/kg and your concentrations is in ng/ml, then dscale = 1e6

NCA_options

specify a list of options for PKNCA to overwrite the defaults (default NULL will use defaults). For example if you want to set the maximum extrapolation of AUCinf to 10 half-life half-life of 0.8 you would use: list(max.aucinf.pext=10, min.hl.r.squared=.9)

NCA_min

minimum number of points required to perform NCA for a given subset (default 4)

analysis_name

string containing the name of the analysis (default 'analysis') to archive to files and reference results later

dsfilter

list of names corresponding to the column names in the dataset and values are a sequence indicating values to keep (default NULL. Multiple names are and-ed together. For example the following would keep all of the records where dose is 1, 2, or 5 and the dose_number is 1

 dsfilter = list(dose=c(1,2,5), dose_number = c(1))

extrap_C0

Boolean variable to enable automatic determination of initial drug concentration if no value is specified; the rules used by WinNonlin will be used:

If the route is "iv infusion" or "extra-vascular" and the data is single dose data, then a concentration of zero will be used. If repeat dosing is used, the minimum value from the previous dosing interval will be used.

If the route is "iv bolus" then log-linear regression of the number of observations specified by extrap_N will be used. If the slope of these points is positive the first positive observation will be used as an estimate of C0

extrap_N

number of points to use for back extrapolation (default 2); this number can be overwritten for each subject using the BACKEXTRAP column in the dataset

sparse

Boolean variable used to indicate data used sparse sampling and the analysis should use the average at each time point (the SPARSEGROUP column must be specified in the dsmap below)

dsmap

list with names specifying the columns in the dataset (* required):

TIME* Time since the first dose; "TIME" (default)

NTIME* Nominal time since last dose; "NTIME" (default)

CONC* Concentration data; "CONC" (default)

DOSE* Dose given; ("DOSE" (default)

ID* Subject ID; ("ID" (default)

ROUTE* Route of administration; "ROUTE" (default), can be either "iv bolus", "iv infusion" or "extra-vascular". Variants such as "IV_bolus" and "extravascular" should work as well.

DOSENUM Numeric dose (starting at 1) used for grouping multiple dose data; optional, NULL (default) for single dose data)

BACKEXTRAP Specifying the number of points to use to extrapolate the initial concentration for "iv bolus" dosing; optoinal f NULL (default) will use the value defined in extrap_N (note this value must be <= NCA_min)

SPARSEGROUP Column containing a unique value grouping cohorts for pooling data. Needed when sparse is set to TRUE; optional, NULL (default)

dsinc

(NOT CURRENTLY IMPLEMENTED) optional character vector of columns from the dataset to include in the output summary (default NULL)

Value

cfg ubiquity system object with the NCA results and if the analysis name is specified:

output/analysis_name-nca_summary-pknca.csv NCA summary

output/analysis_name-pknca_summary.csv Raw output from PKNCA with subject and dose number columns appended

output/analysis_name-nca_data.RData objects containing the NCA summary and a list with the ggplot grobs

See also