Create New system.txt File

Copy a blank template (system_file="template") file to the working directory or an example by specifying the following:

"template" - Empty system file template

"adapt" - Parent/metabolite model taken from the adapt manual used in estimation examples [ADAPT]

"two_cmt_cl" - Two compartment model parameterized in terms of clearances

"one_cmt_cl" - One compartment model parameterized in terms of clearances

"two_cmt_micro" - Two compartment model parameterized in terms of rates (micro constants)

"one_cmt_micro" - One compartment model parameterized in terms of rates (micro constants)

"mab_pk" - General compartmental model of mAb PK from Davda 2014 [DG]

"pbpk" - PBPK model of mAb disposition in mice from Shah 2012 [SB]

"pbpk_template" - System parameters from Shah 2012 [SB] have been defined for all species along with the set notation to be used as a template for developing models with physiological parameters

"pwc" - Example showing how to make if/then or piece-wise continuous variables

"tmdd" - Model of antibody with target-mediated drug disposition

"tumor" - Transit tumor growth model taken from Lobo 2002 [LB]

Usage

system_new(
  file_name = "system.txt",
  system_file = "template",
  overwrite = FALSE,
  output_directory = getwd()
)

Arguments

file_name

name of the new file to create

system_file

name of the system file to copy

overwrite

if TRUE the new system file will overwrite any existing files present

output_directory

getwd() directory where system file will be placed

Value

TRUE if the new file was created and FALSE otherwise

Details

References

[ADAPT] Adapt 5 Users Guide https://bmsr.usc.edu/files/2013/02/ADAPT5-User-Guide.pdf

[DG] Davda et. al. mAbs (2014) 6(4):1094-1102 doi:10.4161/mabs.29095

[LB] Lobo, E.D. & Balthasar, J.P. AAPS J (2002) 4, 212-222 doi:10.1208/ps040442

[SB] Shah, D.K. & Betts, A.M. JPKPD (2012) 39 (1), 67-86 doi:10.1007/s10928-011-9232-2

Examples

# \donttest{
# To create an example system file named example_system.txt:
system_new(system_file      = "mab_pk", 
           file_name        = "system_example.txt", 
           overwrite        = TRUE,  
           output_directory = tempdir())
#> [1] TRUE
# }